# 12_rest/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Quotes API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1.0/quotes": {
            "get": {
                "description": "Returns random quotes",
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "description": "Create a new quote",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Quote data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Quote"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Quote successfully created"
                    }
                }
            }
        },
        "/api/v1.0/quotes/{id}": {
            "get": {
                "description": "Returns a quote by ID",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the quote to retrieve",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "definitions": {
        "Quote": {
            "type": "object",
            "required": ["author", "quote"],
            "properties": {
                "author": {
                    "type": "string",
                    "example": "Steve Jobs"
                },
                "quote": {
                    "type": "string",
                    "example": "Innovation distinguishes between a leader and a follower."
                }
            }
        }
    }
}